Move options types into namespaces
