+++ packages/apollo-angular/src/deprecated.ts
@@ -1,0 +1,15 @@
+import type { OperationVariables } from '@apollo/client';
+import type { Apollo } from './apollo';
+import type { EmptyObject } from './types';
+
+/** @deprecated Use `Apollo.WatchQueryOptions` instead */
+export type WatchQueryOptions<
+  TData = unknown,
+  TVariables extends OperationVariables = EmptyObject,
+> = Apollo.WatchQueryOptions<TData, TVariables>;
+
+/** @deprecated Use `Apollo.MutateOptions` instead */
+export type MutationOptions<
+  TData = unknown,
+  TVariables extends OperationVariables = EmptyObject,
+> = Apollo.MutateOptions<TData, TVariables>;
